Refactor realtime message handling and enhance query validation tests
